Fix bug when getting a random neighbour.
A 0x0 was ambiguous in definition of a candle's neighbours, representing both the lack of a neighbour and the candle at index 0.
Use the candle's own index to represent the lack of a neighbour.
e.g. if the neighbours of candle 4 are given by 0x4436, this means its neighbours are candle 3 and candle 6.